+++ fixture/babel.config.js
@@ -1,7 +1,6 @@
 module.exports = {
   presets: ["module:metro-react-native-babel-preset"],
   plugins: [
-    "react-native-reanimated/plugin",
     [
       "module-resolver",
       {
@@ -13,5 +12,7 @@ module.exports = {
         },
       },
     ],
+    // Reanimated plugin has to be listed last
+    "react-native-reanimated/plugin",
   ],
 };